[breadboard-ui] Move and tidy files a bit
