+++ packages/breadboard-ui/src/index.ts
@@ -4,32 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { BoardList } from "./board-list.js";
-export { Done } from "./done.js";
-export { DrawableInput } from "./drawable.js";
-export { ErrorMessage } from "./error.js";
-export { HistoryTree } from "./history-tree.js";
-export { Input } from "./input.js";
-export { InputContainer } from "./input-container.js";
-export { InputList } from "./input-list.js";
-export { JSONTree } from "./json-tree.js";
-export { Load } from "./load.js";
-export { Output } from "./output.js";
-export { OutputList } from "./output-list.js";
-export { Progress } from "./progress.js";
-export { Start } from "./start.js";
-export { Toast } from "./toast.js";
-export { TraversalControls } from "./traversal-controls.js";
-export { TimelineControls } from "./timeline-controls.js";
-export { UI } from "./ui.js";
-export { WebcamInput } from "./webcam.js";
-
+// TODO: Migrate these to Lit Element and remove them from here.
 import {
   MultipartInput,
   MultipartInputImage,
   MultipartInputText,
-} from "./input-multipart.js";
-import { Diagram } from "./diagram.js";
+} from "./elements/input/input-multipart/input-multipart.js";
+import { Diagram } from "./elements/diagram/diagram.js";
 
 export const register = () => {
   customElements.define("bb-diagram", Diagram);
@@ -38,11 +19,6 @@ export const register = () => {
   customElements.define("bb-multipart-input-text", MultipartInputText);
 };
 
-export type { LoadArgs } from "./load.js";
-export type { OutputArgs } from "./output.js";
-export type { StartArgs } from "./start.js";
-export type { InputArgs } from "./types.js";
-export * as Types from "./types.js";
-export * as Events from "./events.js";
-
-export { StartEvent, ToastEvent, DelayEvent } from "./events.js";
+export * as Types from "./types/types.js";
+export * as Events from "./events/events.js";
+export * as Elements from "./elements/elements.js";
